Include GameBoard#hasWinningPattern
Because:
Either we are playing the game in console or via browser,
we need something that tell us whether there is a winner
or not. Besides that, from outside of the board we probably
care more about whether there is winner or not, rather than
how the winner win.

This commit:
* Implement GameBoard#hasWinningPattern
* Implement GameBoard#hasHorizontalWinningPattern (private)
* Implement GameBoard#hasVerticalWinningPattern (private)
* Implement GameBoard#hasDiagonalWinningPattern (private)
* Include respective test cases